requirements.txt
openai>=1.0.0
chromadb>=0.4.0
langchain>=0.1.0
sentence-transformers>=2.2.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
requests>=2.25.0
